Do not snap map position while panning
When the user is manually interacting with the map using gestures it should not automatically snap to the current instruction on pager advance. However it should still snap to the current instruction when the user swipes the view pager.

This is achieved by tracking the scroll state of the pager to distinguish changes due to user swipe vs. programatic changes.