Fixes #283
Instead of waiting for PeliasPlace call to happen, since the location of the tapped poi is known, we can drop the pin at
the appropriate location before pelias call is made.

`showPlaceSearchFeature` only handles setting the SearchResultAdapter pager.

ReverseGeo fallback for pelias place search only overrides the features and ignores the draw!