Fix implicit intent search when app is in background
Ultimately this turned out to be a race condition that exists on some version of Android which keep the application process alive even if there are no running activities.

In this scenario `MainActivity#executeSearch(String)` would be invoked before `onCreateOptionsMenu(Menu)` so `searchView` would still be `null` at the time. This patch saves the query when this happens and submits it after the search view is created.